Add hot reloading
`Elm.hot.reload(newScope)` updates the `Elm` namespace and hot reloads
all running Elm app instances.

The compiler should first check that a hot reload is safe: If anything
has changed in the "deep canonicalized" type of `main`, it is unsafe.
Then the page should be reloaded instead.

It needs support in elm/browser to work.
(And in elm/virtual-dom for `main : Html msg` programs.)